add three smDescribe* funcs
these are snprintf alike so they should be somewhat composable with
users other printf debugging facilities. behaviour is a bit off from
actual snprintf on the return value as the printers early exit on
filling the buffer ... but still return a value greater than was the
capacity, so half right?